If /var/lib/secrets/ssh doesn't exist, go into "dev mode"
* Don't run sshd

* sleep(1) forever

+++ docker/mgmt/docker-entrypoint.sh
@@ -25,12 +25,18 @@ iwait() {
 
 ################################################
 
-sync_keys
-
-/usr/sbin/sshd -D &
-SSHD_PID=$!
-while iwait -t 300 -r /var/lib/secrets/ssh; do
+if [ -d "$SECRETS_DIR" ]; then
     sync_keys
-    kill -HUP $SSHD_PID  # Will fail, and cause the whole script to exit,
-                         # if sshd exited
-done
+
+    /usr/sbin/sshd -D &
+    SSHD_PID=$!
+    while iwait -t 300 -r /var/lib/secrets/ssh; do
+        sync_keys
+        kill -HUP $SSHD_PID  # Will fail, and cause the whole script to exit,
+                             # if sshd exited
+    done
+else
+    # Development mode
+    sleep infinity
+fi
+
